refactor: separate elements and optimize code
