+++ src/components/admin/financial_assets/datagrid/DatagridElements.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+export const BoxTable = styled.div`
+  height: 600px;
+  width: 100%;
+`
+export const SImg = styled.img`
+  width: 60px;
+  height: 45px;
+`
