+++ src/components/admin/financial_assets/datagrid/index.js
@@ -1,14 +1,14 @@
 import React from 'react'
-import styled from 'styled-components'
 import { DataGrid } from '@material-ui/data-grid'
 import { FiTrash2, FiEdit } from 'react-icons/fi'
 import { IconButton } from '@material-ui/core'
-import Loading from  '../../loading/index'
+import Loading from  '../../../loading/index'
+import { BoxTable, SImg } from './DatagridElements'
 
 const DataList = ({ data, modal, loading }) => {
 
   const thumb = ({ formattedValue }) => {
-    return <img src={formattedValue} className="formatImage" />
+    return <SImg src={formattedValue} />
   }
 
   const actionEdit = ({ id, row }) => {
@@ -105,8 +105,3 @@ const DataList = ({ data, modal, loading }) => {
 }
 
 export default DataList
-
-const BoxTable = styled.div`
-  height: 600px;
-  width: 100%;
-`
