+++ src/components/admin/financial_assets/form/FormElements.js
@@ -1,0 +1,22 @@
+import styled from 'styled-components'
+import { Paper } from '@material-ui/core'
+
+export const Box = styled(Paper)`
+  padding: 25px;
+`
+export const Image = styled.img`
+  max-width: 170px;
+  max-height: 170px;
+  margin: 10px;
+  border: thin solid #eee;
+  border-radius: 5%;
+  overflow: hidden;
+  object-fit: cover;
+`
+
+export const Submit = styled.div`
+  margin: ${({ theme: t }) => t.spacing(0.5)};
+  .buttonSubmit {
+    margin: ${({ theme: t }) => t.spacing(3, 0, 2)};
+  }
+`
